legal_graph/requirements.txt
python-dotenv>=1.0.0
neo4j>=5.14.0
openai>=1.6.0
loguru>=0.7.2
pydantic>=2.0.0
spacy>=3.7.2
fastapi>=0.109.0
uvicorn>=0.27.0
python-multipart>=0.0.6
PyPDF2>=3.0.0
markdown>=3.5.1
python-magic>=0.4.27
